backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0

# Image Processing
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3

# ML/AI
mediapipe==0.10.9
rembg==2.0.56

# Environment
python-dotenv==1.0.0
